Refactor: move AuthService application layers to separate folders
