--- AuthService/Authentication/AuthOperations/IAuthOperations.cs
@@ -1,11 +1,0 @@
-﻿using AuthServiceModelLibrary.ApplicationUser;
-using AuthServiceModelLibrary.DTOs;
-
-namespace AuthService.Authentication.AuthOperations
-{
-    public interface IAuthOperations
-    {
-        bool Register(ApplicationUser user);
-        ApplicationUser VerifyLoginDTO(IUserLoginDTO loginDTO);
-    }
-}
