--- AuthService/Authentication/JWTService/IJWTService.cs
@@ -1,9 +1,0 @@
-﻿using System.IdentityModel.Tokens.Jwt;
-using AuthServiceModelLibrary.ApplicationUser;
-
-namespace AuthService.Authentication.JWTService;
-
-public interface IJWTService
-{
-    JwtSecurityToken GenerateLoginJWT(ApplicationUser user);
-}
