--- AuthService/Authentication/Roles/UserRoles.cs
@@ -1,9 +1,0 @@
-﻿namespace AuthService.Authentication.Roles
-{
-    public enum UserRoles
-    {
-        Administrator,
-        Owner,
-        Tenant
-    }
-}
