--- AuthService/Authentication/Roles/Validator/IRoleValidator.cs
@@ -1,6 +1,0 @@
-﻿namespace AuthService.Authentication.Roles.Validator;
-
-public interface IRoleValidator
-{
-    Task<bool> Validate(string roleName);
-}
