--- AuthService/Authentication/SecurityUtil/SecurityUtil.cs
@@ -1,22 +1,0 @@
-﻿using System.Web.Helpers;
-
-namespace AuthService.Authentication.SecurityUtil
-{
-    public class SecurityUtil : ISecurityUtil
-    {
-        public string CreateSalt()
-        {
-            return Crypto.GenerateSalt();
-        }
-
-        public string HashPassword(string password, string salt)
-        {
-            return Crypto.HashPassword(password + salt);
-        }
-
-        public bool VerifyPassword(string password, string salt, string hashedPassword)
-        {
-            return Crypto.VerifyHashedPassword(hashedPassword, password + salt);
-        }
-    }
-}
