--- AuthService/ModelConverter/IApplicationUserFactory.cs
@@ -1,9 +1,0 @@
-﻿using AuthServiceModelLibrary.ApplicationUser;
-using AuthServiceModelLibrary.DTOs;
-
-namespace AuthService.ModelConverter;
-
-public interface IApplicationUserFactory
-{
-    ApplicationUser Converter(UserRegistrationDTO userDTO);
-}
