+++ AuthService/Program.cs
@@ -1,14 +1,14 @@
 using System.Text;
-using AuthService.Authentication.AuthOperations;
-using AuthService.Authentication.JWTService;
-using AuthService.Authentication.Roles;
-using AuthService.Authentication.Roles.Validator;
-using AuthService.Authentication.SecurityUtil;
-using AuthService.Controller;
-using AuthService.DataAccess;
-using AuthService.DataAccess.UserTableQueries;
-using AuthService.DataSeed;
-using AuthService.ModelConverter;
+using AuthServiceServiceLayer.Authentication.AuthOperations;
+using AuthServiceServiceLayer.Authentication.JWTService;
+using AuthServiceServiceLayer.Authentication.Roles;
+using AuthServiceServiceLayer.Authentication.Roles.Validator;
+using AuthServiceServiceLayer.Authentication.SecurityUtil;
+using AuthServiceAPI.Controller;
+using AuthServiceDataAccess;
+using AuthServiceDataAccess.UserTableQueries;
+using AuthServiceAPI.DataSeed;
+using AuthServiceServiceLayer.ModelConverter;
 using AuthServiceModelLibrary.ApplicationUser;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Identity;
