+++ AuthServiceDataAccess/appsettings.json
@@ -1,0 +1,12 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "ConnString": "Server=(localdb)\\MSSQLLocalDB;Database=PPMAuth;Trusted_Connection=True;"
+  }
+}
