+++ AuthServiceServiceLayer/Authentication/JWTService/IJWTService.cs
@@ -1,0 +1,9 @@
+﻿using System.IdentityModel.Tokens.Jwt;
+using AuthServiceModelLibrary.ApplicationUser;
+
+namespace AuthServiceServiceLayer.Authentication.JWTService;
+
+public interface IJWTService
+{
+    JwtSecurityToken GenerateLoginJWT(ApplicationUser user);
+}
