+++ AuthServiceServiceLayer/Authentication/Roles/UserRoles.cs
@@ -1,0 +1,9 @@
+﻿namespace AuthServiceServiceLayer.Authentication.Roles
+{
+    public enum UserRoles
+    {
+        Administrator,
+        Owner,
+        Tenant
+    }
+}
