+++ AuthServiceServiceLayer/Authentication/Roles/Validator/IRoleValidator.cs
@@ -1,0 +1,6 @@
+﻿namespace AuthServiceServiceLayer.Authentication.Roles.Validator;
+
+public interface IRoleValidator
+{
+    Task<bool> Validate(string roleName);
+}
