+++ AuthServiceServiceLayer/Authentication/SecurityUtil/ISecurityUtil.cs
@@ -1,0 +1,8 @@
+﻿namespace AuthServiceServiceLayer.Authentication.SecurityUtil;
+
+public interface ISecurityUtil
+{
+    string CreateSalt();
+    string HashPassword(string password, string salt);
+    bool VerifyPassword(string password, string salt, string hashedPassword);
+}
