+++ AuthServiceServiceLayer/ModelConverter/IApplicationUserFactory.cs
@@ -1,0 +1,9 @@
+﻿using AuthServiceModelLibrary.ApplicationUser;
+using AuthServiceModelLibrary.DTOs;
+
+namespace AuthServiceServiceLayer.ModelConverter;
+
+public interface IApplicationUserFactory
+{
+    ApplicationUser Converter(UserRegistrationDTO userDTO);
+}
