+++ AuthServiceTests/AuthServiceTests.csproj
@@ -25,7 +25,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AuthService\AuthService.csproj" />
+    <ProjectReference Include="..\AuthService\AuthServiceAPI.csproj" />
   </ItemGroup>
 
 </Project>
